Temp: Comment out column trigger which was causing control to grow after state change and cause infinite layout cycle
i.e. when the break point was reached (e.g. control size 560), VSM was changing the layout and now control size would be larger again (e.g. 605), which would then turn off trigger and re-layout back to old size which would now be smaller again, etc...

We need to ensure the transition between the Right and Vertical states doesn't increase the width of the control.