Switch SettingsExpander to use ItemsRepeater and inherit from Control instead of ItemsControl
This is the WinUI approach TabView/NavigationView uses and allows us to specify the entire Container in the DataTemplate.
This is needed for the ItemSource scenario of SettingsExpander to work properly.
We could also inherit from ContentControl instead of Control like NavigationView, but this was a simpler first step.
Some ItemsControl properties/methods are still missing, but initial approach seems to work and resolve issues with our same desired API surface.
New issue: Expander is shifting position when expanding for some reason now? (don't think it did that before)