Setup ResizeThumb control for ResizeElementAdorner, almost works...
Having separate ResizeThumb which handles manipulation is basically a generalized ContentSizer, this helps encapsulate that logic as well nicely.
Then the ResizeElementAdorner is just a coordinator/initializer to provide the complete package of resizing.
TODO: Moving the element probably can just be handled by the underlying element with standared manipulation (do we have a Behavior for that?) and doesn't need to involve the Adorner layer (outside of syncing movement/position, which needs to be tested...)